Remove [Auth] model that is not related to database
